restrict edit, delete, dashboard to current user

+++ mysite/myapp/views.py
@@ -76,7 +76,9 @@ def create_product(request):
     if request.method == 'POST':
         product_form=ProductForm(request.POST,request.FILES)
         if product_form.is_valid():
-            new_product = product_form.save()
+            new_product = product_form.save(commit=False)
+            new_product.seller = request.user
+            new_product.save()
             return redirect('index')
 
 
@@ -85,6 +87,8 @@ def create_product(request):
 
 def product_edit(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
     product_form = ProductForm(request.POST or None, request.FILES or None, instance=product)
     if request.method=='POST':
         if product_form.is_valid():
@@ -95,13 +99,15 @@ def product_edit(request, id):
 
 def product_delete(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
     if request.method == 'POST':
         product.delete()
         return redirect ('index')
     return render(request, 'myapp/delete.html', {'product':product})
 
 def dashboard(request):
-    products = Product.objects.all()
+    products = Product.objects.filter(seller = request.user) 
     return render(request, 'myapp/dashboard.html', {"products":products})
 
 def register(request):
@@ -114,4 +120,5 @@ def register(request):
     user_form = UserRegistrationForm()
     return render(request,'myapp/register.html',{'user_form':user_form})
     
-
+def invalid(request):
+    return render(request,'myapp/invalid.html')
